Refactoring the email sender into some components (send vs. construction)
