+++ app/org/sagebionetworks/bridge/services/SendMailService.java
@@ -1,18 +1,9 @@
 package org.sagebionetworks.bridge.services;
 
-import java.util.List;
-
-import org.sagebionetworks.bridge.dao.ParticipantOption.SharingScope;
-import org.sagebionetworks.bridge.models.User;
-import org.sagebionetworks.bridge.models.studies.ConsentSignature;
-import org.sagebionetworks.bridge.models.studies.Study;
-import org.sagebionetworks.bridge.models.studies.StudyConsent;
-import org.sagebionetworks.bridge.models.studies.StudyParticipant;
+import org.sagebionetworks.bridge.services.email.MimeTypeEmailProvider;
 
 public interface SendMailService {
 
-    void sendConsentAgreement(User caller, ConsentSignature consent, StudyConsent studyConsent,
-            SharingScope sharingScope);
+    public void sendEmail(MimeTypeEmailProvider provider);
 
-    void sendStudyParticipantsRoster(Study study, List<StudyParticipant> participants);
 }
