+++ app/org/sagebionetworks/bridge/services/email/MimeTypeEmailProvider.java
@@ -1,0 +1,9 @@
+package org.sagebionetworks.bridge.services.email;
+
+import javax.mail.MessagingException;
+
+public interface MimeTypeEmailProvider {
+
+    public MimeTypeEmail getEmail(String defaultFromEmail) throws MessagingException;
+    
+}
