+++ test/org/sagebionetworks/bridge/services/SendMailViaAmazonServiceParticipantRosterTest.java
@@ -15,6 +15,7 @@ import org.mockito.ArgumentCaptor;
 import org.sagebionetworks.bridge.dynamodb.DynamoStudy;
 import org.sagebionetworks.bridge.models.studies.Study;
 import org.sagebionetworks.bridge.models.studies.StudyParticipant;
+import org.sagebionetworks.bridge.services.email.ParticipantRosterProvider;
 
 import com.amazonaws.services.simpleemail.AmazonSimpleEmailServiceClient;
 import com.amazonaws.services.simpleemail.model.SendRawEmailRequest;
@@ -47,87 +48,6 @@ public class SendMailViaAmazonServiceParticipantRosterTest {
     }
     
     @Test
-    public void participantsCorrectlyDescribedInText() {
-        StudyParticipant participant = new StudyParticipant();
-        participant.setFirstName("First");
-        participant.setLastName("Last");
-        participant.setEmail("test@test.com");
-        participant.setPhone("(123) 456-7890");
-        participant.put("recontact", "true");
-        List<StudyParticipant> participants = Lists.newArrayList(participant);
-
-        String output = service.createInlineParticipantRoster(study, participants);
-        
-        assertEquals("There is 1 user enrolled in this study:\n\ntest@test.com (First Last)\nPhone: (123) 456-7890\nRecontact: true\n", output);
-        
-        participant.setLastName(null);
-        output = service.createInlineParticipantRoster(study, participants);
-        assertEquals("There is 1 user enrolled in this study:\n\ntest@test.com (First)\nPhone: (123) 456-7890\nRecontact: true\n", output);
-        
-        participant.setFirstName(null);
-        participant.setLastName("Last");
-        output = service.createInlineParticipantRoster(study, participants);
-        assertEquals("There is 1 user enrolled in this study:\n\ntest@test.com (Last)\nPhone: (123) 456-7890\nRecontact: true\n", output);
-        
-        participant.setPhone(null);
-        output = service.createInlineParticipantRoster(study, participants);
-        assertEquals("There is 1 user enrolled in this study:\n\ntest@test.com (Last)\nPhone: \nRecontact: true\n", output);
-        
-        participant.remove("recontact");
-        output = service.createInlineParticipantRoster(study, participants);
-        assertEquals("There is 1 user enrolled in this study:\n\ntest@test.com (Last)\nPhone: \nRecontact: \n", output);
-        
-        StudyParticipant numberTwo = new StudyParticipant();
-        numberTwo.setEmail("test2@test.com");
-        
-        participants.add(numberTwo);
-        output = service.createInlineParticipantRoster(study, participants);
-        assertTrue(output.contains("There are 2 users enrolled in this study:"));
-        
-        participants.clear();
-        output = service.createInlineParticipantRoster(study, participants);
-        assertTrue(output.contains("There are no users enrolled in this study."));
-    }
-    
-    @Test
-    public void participantsCorrectlyDescribedInCSV() {
-        StudyParticipant participant = new StudyParticipant();
-        participant.setFirstName("First");
-        participant.setLastName("Last");
-        participant.setEmail("test@test.com");
-        // Tab snuck into this string should be converted to a space
-        participant.setPhone("(123)\t456-7890");
-        List<StudyParticipant> participants = Lists.newArrayList(participant);
-
-        String output = service.createParticipantCSV(study, participants);
-        assertEquals("Email\tFirst Name\tLast Name\tPhone\tRecontact\ntest@test.com\tFirst\tLast\t(123) 456-7890\t\n", output);
-        
-        participant.setLastName(null);
-        output = service.createParticipantCSV(study, participants);
-        assertEquals("Email\tFirst Name\tLast Name\tPhone\tRecontact\ntest@test.com\tFirst\t\t(123) 456-7890\t\n", output);
-        
-        participant.setFirstName(null);
-        participant.setLastName("Last");
-        output = service.createParticipantCSV(study, participants);
-        assertEquals("Email\tFirst Name\tLast Name\tPhone\tRecontact\ntest@test.com\t\tLast\t(123) 456-7890\t\n", output);
-        
-        participant.setPhone(null);
-        output = service.createParticipantCSV(study, participants);
-        assertEquals("Email\tFirst Name\tLast Name\tPhone\tRecontact\ntest@test.com\t\tLast\t\t\n", output);
-        
-        StudyParticipant numberTwo = new StudyParticipant();
-        numberTwo.setEmail("test2@test.com");
-        
-        participants.add(numberTwo);
-        output = service.createParticipantCSV(study, participants);
-        assertEquals("Email\tFirst Name\tLast Name\tPhone\tRecontact\ntest@test.com\t\tLast\t\t\ntest2@test.com\t\t\t\t\n", output);
-        
-        participants.clear();
-        output = service.createParticipantCSV(study, participants);
-        assertEquals("Email\tFirst Name\tLast Name\tPhone\tRecontact\n", output);
-    }
-    
-    @Test
     public void sendParticipantRoster() {
         StudyParticipant participant = new StudyParticipant();
         participant.setFirstName("First");
@@ -136,7 +56,8 @@ public class SendMailViaAmazonServiceParticipantRosterTest {
         participant.setPhone("(123) 456-7890");
         List<StudyParticipant> participants = Lists.newArrayList(participant);
         
-        service.sendStudyParticipantsRoster(study, participants);
+        ParticipantRosterProvider provider = new ParticipantRosterProvider(study, participants);
+        service.sendEmail(provider);
         
         verify(emailClient).sendRawEmail(argument.capture());
