+++ test/org/sagebionetworks/bridge/services/email/ConsentEmailProviderTest.java
@@ -1,0 +1,5 @@
+package org.sagebionetworks.bridge.services.email;
+
+public class ConsentEmailProviderTest {
+
+}
